requirements.txt
pytesseract
Pillow
requests
numpy
beautifulsoup4


